extracted the Events On Demand box
