+++ composables/useEventsOnDemand.ts
@@ -1,0 +1,5 @@
+const useEventsOnDemand = () => {
+      return useLazyFetch("/api/EventsOnDemand")
+}
+
+export default useEventsOnDemand;
